Update FastAPI project structure and settings
This update makes several key changes. It includes the introduction of specific package versions in the requirements.txt file, and the creation of a new .gitignore file. Additionally, SSL certificates are integrated in the Nginix Docker setting and Nginix configuration files are revised accordingly. The FastAPI application is also extended to support Server-Sent Events and OpenAPI documentation customized. Lastly, code comments and documentation including README files are updated for clarity and to reflect the changes made.